split openshift-resources in 3
