+++ reconcile/openshift_routes.py
@@ -1,0 +1,19 @@
+import semver
+
+import reconcile.openshift_resources_base as orb
+
+QONTRACT_INTEGRATION = 'openshift-routes'
+QONTRACT_INTEGRATION_VERSION = semver.format_version(1, 9, 3)
+
+
+def run(dry_run=False, thread_pool_size=10, internal=None,
+        use_jump_host=True, defer=None):
+    providers = ['route']
+    orb.QONTRACT_INTEGRATION = QONTRACT_INTEGRATION
+    orb.QONTRACT_INTEGRATION_VERSION = QONTRACT_INTEGRATION_VERSION
+
+    orb.run(dry_run=dry_run,
+            thread_pool_size=thread_pool_size,
+            internal=internal,
+            use_jump_host=use_jump_host,
+            providers=providers)
